backend/requirements.txt
django
djangorestframework
django-environ
psycopg2
djangorestframework_simplejwt
requests
lxml
six
django-storages
boto3
django-cors-headers
Pillow
drf-writable-nested